requirements.txt
certifi==2016.8.31
elasticsearch==2.4.0
future==0.15.2
nltk==3.2.1
oauthlib==2.0.0
python-telegram-bot==5.0.0
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
textblob==0.11.1
tweepy==3.5.0
urllib3==1.17
